+++ R/read_excel.R
@@ -14,8 +14,10 @@ NULL
 #'   unskipped column.
 #' @param col_types Either `NULL` to guess from the spreadsheet or a character
 #'   vector containing one entry per column from these options: "skip",
-#'   "numeric", "date" or "text". The content of a cell in a skipped column is
-#'   never read and that column will not appear in the data frame output.
+#'   "numeric", "date", "text" or "list". The content of a cell in a skipped column is
+#'   never read and that column will not appear in the data frame output. A list cell
+#'   loads a column as a list of length 1 vectors, which are typed using the type
+#'   guessing logic from `col_types = NULL`, but on a cell-by-cell basis.
 #' @param na Character vector of strings to use for missing values. By default,
 #'   readxl treats blank cells as missing data.
 #' @param skip Number of rows to skip before reading any data. Leading blank
@@ -146,7 +148,7 @@ check_col_types <- function(col_types) {
     col_types[blank] <- "skip"
   }
 
-  accepted_types <- c("skip", "numeric", "date", "text")
+  accepted_types <- c("skip", "numeric", "date", "text", "list")
   ok <- col_types %in% accepted_types
   if (any(!ok)) {
     info <- paste(