+++ src/XlsWorkSheet.h
@@ -67,9 +67,9 @@ public:
     return out;
   }
 
-  std::vector<CellType> colTypes(const StringSet &na, int nskip = 0,
+  std::vector<ColType> colTypes(const StringSet &na, int nskip = 0,
                                  int guess_max = 1000) {
-    std::vector<CellType> types(ncol_);
+    std::vector<ColType> types(ncol_);
 
     for (int i = nskip; i < nrow_ && i < nskip + guess_max; ++i) {
       if ((i + 1) % 10000 == 0)
@@ -78,7 +78,7 @@ public:
       xls::st_row::st_row_data row = pWS_->rows.row[i];
 
       for (int j = 0; j < ncol_; ++j) {
-        CellType type = cellType(row.cells.cell[j], &pWS_->workbook->xfs, customDateFormats_, na);
+        ColType type = as_ColType(cellType(row.cells.cell[j], &pWS_->workbook->xfs, customDateFormats_, na));
 
         // Excel is simple enough we can enforce a strict ordering
         if (type > types[j]) {
@@ -90,7 +90,7 @@ public:
     return types;
   }
 
-  Rcpp::List readCols(Rcpp::CharacterVector names, std::vector<CellType> types,
+  Rcpp::List readCols(Rcpp::CharacterVector names, std::vector<ColType> types,
                       const StringSet &na, int nskip = 0) {
     if ((int) names.size() != ncol_ || (int) types.size() != ncol_){
       Rcpp::stop("Received %d names and %d types, but worksheet contains %d columns.",
@@ -117,29 +117,26 @@ public:
 
         // Needs to compare to actual cell type to give warnings
         switch(types[j]) {
-        case CELL_BLANK:
+        case COL_BLANK:
+        case COL_SKIP:
           break;
-        case CELL_SKIP:
-          break;
-        case CELL_NUMERIC:
+        case COL_NUMERIC:
           switch(type) {
           case CELL_BLANK:
             REAL(col)[i] = NA_REAL;
             break;
           case CELL_NUMERIC:
-          case CELL_SKIP:
           case CELL_DATE:
             REAL(col)[i] = cell.d;
             break;
-          case CELL_TEXT:
+          case COL_TEXT:
             Rcpp::warning("Expecting numeric in [%i, %i] got `%s`",
               i + 1, j + 1, (char*) cell.str);
             REAL(col)[i] = NA_REAL;
           }
           break;
-        case CELL_DATE:
+        case COL_DATE:
           switch(type) {
-          case CELL_SKIP:
           case CELL_BLANK:
             REAL(col)[i] = NA_REAL;
             break;
@@ -158,7 +155,7 @@ public:
             break;
           }
           break;
-        case CELL_TEXT:
+        case COL_TEXT:
           if (type == CELL_BLANK) {
             SET_STRING_ELT(col, i, NA_STRING);
           } else {
@@ -167,6 +164,30 @@ public:
             SET_STRING_ELT(col, i, rString);
           }
           break;
+        case COL_LIST:
+          switch(type) {
+          case CELL_BLANK: {
+            SET_VECTOR_ELT(col, i, Rf_ScalarLogical(NA_LOGICAL));
+            break;
+          }
+          case CELL_NUMERIC: {
+            SET_VECTOR_ELT(col, i, Rf_ScalarReal(cell.d));
+            break;
+          }
+          case CELL_DATE: {
+            Rcpp::RObject cell_val = Rf_ScalarReal((cell.d - offset_) * 86400);
+            cell_val.attr("class") = Rcpp::CharacterVector::create("POSIXct", "POSIXt");
+            cell_val.attr("tzone") = "UTC";
+            SET_VECTOR_ELT(col, i, cell_val);
+            break;
+          }
+          case CELL_TEXT: {
+            std::string stdString((char*) cell.str);
+            Rcpp::CharacterVector rString = na.contains(stdString) ? NA_STRING : Rf_mkCharCE(stdString.c_str(), CE_UTF8);
+            SET_VECTOR_ELT(col, i, rString);
+            break;
+          }
+          }
         }
       }
     }